+++ druid-windows/src/handle_cli_info_flags.c
@@ -17,9 +17,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <unistd.h>
 #include <stddef.h>
-#include <systemd/sd-device.h>
 #include "druid.h"
 
 /**
@@ -38,11 +36,11 @@
  */
 int handle_cli_info_flags(int ac, char **av)
 {
-    size_t help_file_return_value = display_file(ac, av,
+    long unsigned int help_file_return_value = display_file(ac, av,
         HELP_FLAG, HELP_FLAG_OPTION, HELP_FILE);
-    size_t format_file_return_value = display_file(ac, av,
+    long unsigned int format_file_return_value = display_file(ac, av,
         FORMAT_FLAG, FORMAT_FLAG_OPTION, FORMAT_FILE);
-    size_t license_file_return_value = display_file(ac, av,
+    long unsigned int license_file_return_value = display_file(ac, av,
         LICENSE_FLAG, LICENSE_FLAG_OPTION, LICENSE_FILE);
     
     if (help_file_return_value == SUCCESS)
